+++ amplitude.js
@@ -4948,7 +4948,7 @@ module.exports = uuid;
 
 }, {}],
 10: [function(require, module, exports) {
-module.exports = '3.2.0';
+module.exports = '3.3.0';
 
 }, {}],
 11: [function(require, module, exports) {