+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+### 3.3.0 (October 19, 2016)
+
 * Add option to track GCLID (Google Click ID) as a user property (set `includeGclid` to `true` in the SDK configuration).
 * Add option to track new UTM parameters, referrer, and GCLID values during the same session. By default the SDK only saves the values once at the start of the session. You can remove this restriction by setting `saveParamsReferrerOncePerSession` to `false` in the SDK configuration. See the [Readme](https://github.com/amplitude/Amplitude-Javascript#tracking-utm-parameters-referrer-and-gclid) for more information.